move isGranularityNaturalAligned() to utils to be reused
